Add a Number data type
This adds a new data type that matches a Numeric or any String that
looks like a Numeric.

This is useful when you want to accept a number but don't care if it's
quoted or not. For example when the number will be cast to a String in a
template.

Another solution to this would be to use `Init[Numeric]`
However, that wasn't added until Puppet 5.10 in [PUP-7277](https://tickets.puppetlabs.com/browse/PUP-7277).